+++ packages/core/src/core/handleEvents.ts
@@ -25,7 +25,7 @@ const HandleEvents = {
       console.log('handleHttp data', data);
 
       const result = httpTransform(data);
-      console.log(type);
+      console.log('result', result);
       console.log('====================================');
       // 添加用户行为，去掉自身上报的接口行为
       if (!data.url.includes(options.dsn)) {
@@ -38,6 +38,7 @@ const HandleEvents = {
         });
       }
 
+      console.log('result.status', result.status);
       if (result.status === 'error') {
         // 上报接口错误
         transportData.send({ ...result, type, status: STATUS_CODE.ERROR });