+++ test/src/main.js
@@ -5,13 +5,13 @@ import store from './store';
 import ElementUI from 'element-ui';
 import 'element-ui/lib/theme-chalk/index.css';
 
-// import ZMonitor from '../../packages/core/src';
-// import performance from '../../packages/performance/src';
-// import recordscreen from '../../packages/recordscreen/src';
+import ZMonitor from '../../packages/core/src';
+import performance from '../../packages/performance/src';
+import recordscreen from '../../packages/recordscreen/src';
 
-import ZMonitor from '@zmonitor/core';
-import performance from '@zmonitor/performance';
-import recordscreen from '@zmonitor/recordscreen';
+// import ZMonitor from '@zmonitor/core';
+// import performance from '@zmonitor/performance';
+// import recordscreen from '@zmonitor/recordscreen';
 
 Vue.use(ZMonitor, {
   dsn: 'http://localhost:8080/reportData',
@@ -31,8 +31,6 @@ Vue.use(ZMonitor, {
         };
       }
     }
-
-    console.log('data', data);
     let { url, response } = data;
     // code为200，接口正常，反之亦然
     let { code } = typeof response === 'string' ? safeStringify(response) : response;