+++ packages/recordscreen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zmonitor/recordscreen
 
+## 2.0.1
+
+### Patch Changes
+
+- @zmonitor/utils@2.0.1
+
 ## 2.0.0
 
 ### Major Changes